reorganize collector files
