+++ indexer/src/collectors/github/common.ts
@@ -1,22 +1,22 @@
 import { Repository, In } from "typeorm";
-import { PageInfo } from "../../../events/github/unpaginate.js";
+import { PageInfo } from "../../events/github/unpaginate.js";
 import {
   Artifact,
   Project,
   ArtifactType,
   ArtifactNamespace,
-} from "../../../db/orm-entities.js";
-import { GenericError } from "../../../common/errors.js";
-import { IEventRecorder } from "../../../recorder/types.js";
-import { TimeSeriesCacheWrapper } from "../../../cacher/time-series.js";
+} from "../../db/orm-entities.js";
+import { GenericError } from "../../common/errors.js";
+import { IEventRecorder } from "../../recorder/types.js";
+import { TimeSeriesCacheWrapper } from "../../cacher/time-series.js";
 import { ClientError, RequestDocument, Variables } from "graphql-request";
-import { graphQLClient } from "../../../events/github/graphQLClient.js";
+import { graphQLClient } from "../../events/github/graphQLClient.js";
 import { DateTime } from "luxon";
-import { logger } from "../../../utils/logger.js";
+import { logger } from "../../utils/logger.js";
 import {
   BatchedProjectArtifactsCollector,
   ProjectArtifactsCollector,
-} from "../../../scheduler/common.js";
+} from "../../scheduler/common.js";
 import { Mutex } from "async-mutex";
 
 export class IncompleteRepoName extends GenericError {}
